+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 * Added possibility to save an entity child
+* Added possibility to save many entity child
+* Added possibility to delete child